BLOJSOM-259.
Adds Tomcat Manager tasks to build.xml and support for "QuickrStart" builds from an IDE.

QuickrStart directory includes lib and properties. Adds HSQLDB lib and properties including a "full-initial-data-hsqldb" script. A custom blojsom.xml is included that loads the Hibernate driver for the QuickrStart bundle.

To build the QuickrStart bundle run the quickrstart target in the build.xml file.