blojsom, being the venerable platform it is, has a number of legacy JARs that are not available in Maven Central. Previously, blojsom instructed builders to source JARs from the legacy SVN, or more recently, the svn2git repo and install them locally.
This branch should address this "instructable" and build time error, allowing
for better CI integration as well.

Added a `legacy-jars-repo` `repository` to the Parent POM. Three (3) 
JARs are currently part of that effort, `conten-0.4.1`, `netscape-ldapsdk-4.17`, 
and `webmacro-2.0`. The legacy `textile4j-1.20` JAR has been removed and replaced
with a modern parser.