Various elements of work on Bulk Response Management [skip ci]
Set a admin page comment on all templates to match to plugin and action.

H2/H3 work on bulk-response-management template.

CSS work on semantics of bulk-response-management. With stubbing out of 
accordion work.

Template work on reorganization of page scripting and stubbed out removal of
scriptaculous in preparation of YUI2 event handling. YUI2 conversion to 
pave way for move to YUI3 and Vanilla JS work.