Update log4j
This patches the Log4j version to 2.17.0. Corrected the maven version
range syntax to follow the specification. Take the opportunity to also
tweak the Twitter4J plugin.

Fixed the URL to Apples 2004 press release and the bit.ly URL to the
blojsom archive wiki to use HTTPS.

Increment PATCH version to 3.5.2-SNAPSHOT