Resolve build failure due to local repo lookups
Submodules needing runtime or legacy jars not in Maven Central would not
build when referencing the local repos.

Add-ons and the Quick Start War now have references to repo in the
parent via the same name and should build independent of the parent.

This results in submodules inheriting a repository that may or may not
resolve, but will not be relevant to the project in question. Viewing
the effective-pom via help:effective-pom or your IDE will show this
inheritance.

Some consideration was given to the directory-maven-plugin and the
maven.multiModuleProjectDirectory property--an internal value subject to
removal or change, but this suffices for the moment.